feat | Interface Done

+++ app/src/main/java/mx/kodemia/appclimaedwin/network/Sys.kt
@@ -1,7 +1,7 @@
 package mx.kodemia.appclimaedwin.network
 
 data class Sys(
-    val sunrise: Double,
-    val sunset: Double,
+    val sunrise: Long,
+    val sunset: Long,
     val country: String
 )
